Improve performance by 20x with jsonb querying
Rather than composing a query which contains a huge quantity of IDs in
an IN clause, we can push the id curation logic into a subquery so that
we don't pay the (dramatic, 8-10s) performance penalty of sanitizing
tons of parameters.

The downside is that the query still takes a while to run, about 500ms
on my laptop, so I'm only calling this a 20x performance improvement.
It's likely there is more low hanging fruit here with some proper
indexing of these JSON documents.